Refactor Sage Terminal app structure
