+++ app/terminal/src/app/commands.rs
@@ -1,0 +1,8 @@
+#[path = "commands/dispatch.rs"]
+mod dispatch;
+#[path = "commands/model.rs"]
+mod model;
+#[path = "commands/session.rs"]
+mod session;
+#[path = "commands/skills.rs"]
+mod skills;
