+++ app/terminal/src/app/surfaces/popup.rs
@@ -1,0 +1,6 @@
+#[path = "popup/core.rs"]
+mod core;
+#[path = "popup/provider.rs"]
+mod provider;
+#[path = "popup/skill.rs"]
+mod skill;
