+++ app/terminal/src/custom_terminal.rs
@@ -1,19 +1,18 @@
+#[path = "custom_terminal/diff.rs"]
+mod diff;
+#[path = "custom_terminal/draw.rs"]
+mod draw;
+
 use std::io;
 use std::io::Write;
 
-use crossterm::cursor::MoveTo;
-use crossterm::queue;
-use crossterm::style::{
-    Attribute as CAttribute, Colors, Print, SetAttribute, SetBackgroundColor, SetColors,
-    SetForegroundColor,
-};
-use crossterm::terminal::Clear;
-use ratatui::backend::{Backend, ClearType, CrosstermBackend};
-use ratatui::buffer::{Buffer, Cell};
+use ratatui::backend::{Backend, CrosstermBackend};
+use ratatui::buffer::Buffer;
 use ratatui::layout::{Position, Rect, Size};
-use ratatui::style::{Color, Modifier};
 use ratatui::widgets::Widget;
-use unicode_width::UnicodeWidthStr;
+
+use self::diff::{diff_buffers, viewport_rect};
+use self::draw::draw_commands;
 
 pub type BackendImpl = CrosstermBackend<std::io::Stdout>;
 
@@ -127,7 +126,7 @@ where
     pub fn clear(&mut self) -> io::Result<()> {
         self.backend
             .set_cursor_position(self.viewport_area.as_position())?;
-        self.backend.clear_region(ClearType::AfterCursor)?;
+        self.backend.clear_region(ratatui::backend::ClearType::AfterCursor)?;
         self.previous_buffer_mut().reset();
         Ok(())
     }
@@ -201,8 +200,8 @@ where
     fn flush(&mut self) -> io::Result<()> {
         let updates = diff_buffers(self.previous_buffer(), self.current_buffer());
         if let Some((x, y)) = updates.iter().rev().find_map(|command| match command {
-            DrawCommand::Put { x, y, .. } => Some((*x, *y)),
-            DrawCommand::ClearToEnd { .. } => None,
+            diff::DrawCommand::Put { x, y, .. } => Some((*x, *y)),
+            diff::DrawCommand::ClearToEnd { .. } => None,
         }) {
             self.last_known_cursor_pos = Position { x, y };
         }
@@ -228,208 +227,5 @@ where
     fn swap_buffers(&mut self) {
         self.previous_buffer_mut().reset();
         self.current = 1 - self.current;
-    }
-}
-
-fn viewport_rect(screen_size: Size, viewport_height: u16, anchor_y: u16) -> Rect {
-    let height = viewport_height.max(1).min(screen_size.height.max(1));
-    let max_y = screen_size.height.saturating_sub(height);
-    let y = anchor_y.min(max_y);
-    Rect::new(0, y, screen_size.width, height)
-}
-
-#[derive(Debug)]
-enum DrawCommand {
-    Put { x: u16, y: u16, cell: Cell },
-    ClearToEnd { x: u16, y: u16, bg: Color },
-}
-
-fn diff_buffers(previous: &Buffer, next: &Buffer) -> Vec<DrawCommand> {
-    let mut updates = Vec::new();
-    let mut last_nonblank_columns = vec![0; previous.area.height as usize];
-
-    for y in 0..previous.area.height {
-        let row_start = y as usize * previous.area.width as usize;
-        let row_end = row_start + previous.area.width as usize;
-        let row = &next.content[row_start..row_end];
-        let bg = row.last().map(|cell| cell.bg).unwrap_or(Color::Reset);
-
-        let mut last_nonblank_column = 0usize;
-        let mut column = 0usize;
-        while column < row.len() {
-            let cell = &row[column];
-            let width = display_width(cell.symbol()).max(1);
-            if cell.symbol() != " " || cell.bg != bg || cell.modifier != Modifier::empty() {
-                last_nonblank_column = column + (width.saturating_sub(1));
-            }
-            column += width;
-        }
-
-        if last_nonblank_column + 1 < row.len() {
-            let (x, y) = previous.pos_of(row_start + last_nonblank_column + 1);
-            updates.push(DrawCommand::ClearToEnd { x, y, bg });
-        }
-
-        last_nonblank_columns[y as usize] = last_nonblank_column as u16;
-    }
-
-    let mut invalidated = 0usize;
-    let mut to_skip = 0usize;
-    for (i, (current, prior)) in next.content.iter().zip(previous.content.iter()).enumerate() {
-        if !current.skip && (current != prior || invalidated > 0) && to_skip == 0 {
-            let (x, y) = previous.pos_of(i);
-            let row = i / previous.area.width as usize;
-            if x <= last_nonblank_columns[row] {
-                updates.push(DrawCommand::Put {
-                    x,
-                    y,
-                    cell: current.clone(),
-                });
-            }
-        }
-
-        to_skip = display_width(current.symbol()).saturating_sub(1);
-        let affected_width = std::cmp::max(
-            display_width(current.symbol()),
-            display_width(prior.symbol()),
-        );
-        invalidated = std::cmp::max(affected_width, invalidated).saturating_sub(1);
-    }
-
-    updates
-}
-
-fn draw_commands<I>(writer: &mut impl Write, commands: I) -> io::Result<()>
-where
-    I: Iterator<Item = DrawCommand>,
-{
-    let mut fg = Color::Reset;
-    let mut bg = Color::Reset;
-    let mut modifier = Modifier::empty();
-    let mut last_pos: Option<Position> = None;
-
-    for command in commands {
-        let (x, y) = match &command {
-            DrawCommand::Put { x, y, .. } => (*x, *y),
-            DrawCommand::ClearToEnd { x, y, .. } => (*x, *y),
-        };
-
-        if !matches!(last_pos, Some(pos) if x == pos.x + 1 && y == pos.y) {
-            queue!(writer, MoveTo(x, y))?;
-        }
-        last_pos = Some(Position { x, y });
-
-        match command {
-            DrawCommand::Put { cell, .. } => {
-                if cell.modifier != modifier {
-                    ModifierDiff {
-                        from: modifier,
-                        to: cell.modifier,
-                    }
-                    .queue(writer)?;
-                    modifier = cell.modifier;
-                }
-                if cell.fg != fg || cell.bg != bg {
-                    queue!(
-                        writer,
-                        SetColors(Colors::new(cell.fg.into(), cell.bg.into()))
-                    )?;
-                    fg = cell.fg;
-                    bg = cell.bg;
-                }
-                queue!(writer, Print(cell.symbol()))?;
-            }
-            DrawCommand::ClearToEnd { bg: clear_bg, .. } => {
-                queue!(writer, SetAttribute(CAttribute::Reset))?;
-                modifier = Modifier::empty();
-                queue!(writer, SetBackgroundColor(clear_bg.into()))?;
-                bg = clear_bg;
-                queue!(writer, Clear(crossterm::terminal::ClearType::UntilNewLine))?;
-            }
-        }
-    }
-
-    queue!(
-        writer,
-        SetForegroundColor(crossterm::style::Color::Reset),
-        SetBackgroundColor(crossterm::style::Color::Reset),
-        SetAttribute(CAttribute::Reset)
-    )?;
-    Ok(())
-}
-
-fn display_width(text: &str) -> usize {
-    if !text.contains('\x1B') {
-        return UnicodeWidthStr::width(text);
-    }
-
-    let mut visible = String::with_capacity(text.len());
-    let mut chars = text.chars();
-    while let Some(ch) = chars.next() {
-        if ch == '\x1B' && chars.clone().next() == Some(']') {
-            let _ = chars.next();
-            for c in chars.by_ref() {
-                if c == '\x07' {
-                    break;
-                }
-            }
-            continue;
-        }
-        visible.push(ch);
-    }
-    UnicodeWidthStr::width(visible.as_str())
-}
-
-struct ModifierDiff {
-    from: Modifier,
-    to: Modifier,
-}
-
-impl ModifierDiff {
-    fn queue<W: Write>(self, writer: &mut W) -> io::Result<()> {
-        let removed = self.from - self.to;
-        if removed.contains(Modifier::REVERSED) {
-            queue!(writer, SetAttribute(CAttribute::NoReverse))?;
-        }
-        if removed.contains(Modifier::BOLD) {
-            queue!(writer, SetAttribute(CAttribute::NormalIntensity))?;
-            if self.to.contains(Modifier::DIM) {
-                queue!(writer, SetAttribute(CAttribute::Dim))?;
-            }
-        }
-        if removed.contains(Modifier::ITALIC) {
-            queue!(writer, SetAttribute(CAttribute::NoItalic))?;
-        }
-        if removed.contains(Modifier::UNDERLINED) {
-            queue!(writer, SetAttribute(CAttribute::NoUnderline))?;
-        }
-        if removed.contains(Modifier::DIM) {
-            queue!(writer, SetAttribute(CAttribute::NormalIntensity))?;
-        }
-        if removed.contains(Modifier::CROSSED_OUT) {
-            queue!(writer, SetAttribute(CAttribute::NotCrossedOut))?;
-        }
-
-        let added = self.to - self.from;
-        if added.contains(Modifier::REVERSED) {
-            queue!(writer, SetAttribute(CAttribute::Reverse))?;
-        }
-        if added.contains(Modifier::BOLD) {
-            queue!(writer, SetAttribute(CAttribute::Bold))?;
-        }
-        if added.contains(Modifier::ITALIC) {
-            queue!(writer, SetAttribute(CAttribute::Italic))?;
-        }
-        if added.contains(Modifier::UNDERLINED) {
-            queue!(writer, SetAttribute(CAttribute::Underlined))?;
-        }
-        if added.contains(Modifier::DIM) {
-            queue!(writer, SetAttribute(CAttribute::Dim))?;
-        }
-        if added.contains(Modifier::CROSSED_OUT) {
-            queue!(writer, SetAttribute(CAttribute::CrossedOut))?;
-        }
-
-        Ok(())
     }
 }
