+++ app/terminal/src/custom_terminal/draw.rs
@@ -1,0 +1,127 @@
+use std::io;
+use std::io::Write;
+
+use crossterm::cursor::MoveTo;
+use crossterm::queue;
+use crossterm::style::{
+    Attribute as CAttribute, Colors, Print, SetAttribute, SetBackgroundColor, SetColors,
+    SetForegroundColor,
+};
+use crossterm::terminal::Clear;
+use ratatui::layout::Position;
+use ratatui::style::{Color, Modifier};
+
+use super::diff::DrawCommand;
+
+pub(super) fn draw_commands<I>(writer: &mut impl Write, commands: I) -> io::Result<()>
+where
+    I: Iterator<Item = DrawCommand>,
+{
+    let mut fg = Color::Reset;
+    let mut bg = Color::Reset;
+    let mut modifier = Modifier::empty();
+    let mut last_pos: Option<Position> = None;
+
+    for command in commands {
+        let (x, y) = match &command {
+            DrawCommand::Put { x, y, .. } => (*x, *y),
+            DrawCommand::ClearToEnd { x, y, .. } => (*x, *y),
+        };
+
+        if !matches!(last_pos, Some(pos) if x == pos.x + 1 && y == pos.y) {
+            queue!(writer, MoveTo(x, y))?;
+        }
+        last_pos = Some(Position { x, y });
+
+        match command {
+            DrawCommand::Put { cell, .. } => {
+                if cell.modifier != modifier {
+                    ModifierDiff {
+                        from: modifier,
+                        to: cell.modifier,
+                    }
+                    .queue(writer)?;
+                    modifier = cell.modifier;
+                }
+                if cell.fg != fg || cell.bg != bg {
+                    queue!(
+                        writer,
+                        SetColors(Colors::new(cell.fg.into(), cell.bg.into()))
+                    )?;
+                    fg = cell.fg;
+                    bg = cell.bg;
+                }
+                queue!(writer, Print(cell.symbol()))?;
+            }
+            DrawCommand::ClearToEnd { bg: clear_bg, .. } => {
+                queue!(writer, SetAttribute(CAttribute::Reset))?;
+                modifier = Modifier::empty();
+                queue!(writer, SetBackgroundColor(clear_bg.into()))?;
+                bg = clear_bg;
+                queue!(writer, Clear(crossterm::terminal::ClearType::UntilNewLine))?;
+            }
+        }
+    }
+
+    queue!(
+        writer,
+        SetForegroundColor(crossterm::style::Color::Reset),
+        SetBackgroundColor(crossterm::style::Color::Reset),
+        SetAttribute(CAttribute::Reset)
+    )?;
+    Ok(())
+}
+
+struct ModifierDiff {
+    from: Modifier,
+    to: Modifier,
+}
+
+impl ModifierDiff {
+    fn queue<W: Write>(self, writer: &mut W) -> io::Result<()> {
+        let removed = self.from - self.to;
+        if removed.contains(Modifier::REVERSED) {
+            queue!(writer, SetAttribute(CAttribute::NoReverse))?;
+        }
+        if removed.contains(Modifier::BOLD) {
+            queue!(writer, SetAttribute(CAttribute::NormalIntensity))?;
+            if self.to.contains(Modifier::DIM) {
+                queue!(writer, SetAttribute(CAttribute::Dim))?;
+            }
+        }
+        if removed.contains(Modifier::ITALIC) {
+            queue!(writer, SetAttribute(CAttribute::NoItalic))?;
+        }
+        if removed.contains(Modifier::UNDERLINED) {
+            queue!(writer, SetAttribute(CAttribute::NoUnderline))?;
+        }
+        if removed.contains(Modifier::DIM) {
+            queue!(writer, SetAttribute(CAttribute::NormalIntensity))?;
+        }
+        if removed.contains(Modifier::CROSSED_OUT) {
+            queue!(writer, SetAttribute(CAttribute::NotCrossedOut))?;
+        }
+
+        let added = self.to - self.from;
+        if added.contains(Modifier::REVERSED) {
+            queue!(writer, SetAttribute(CAttribute::Reverse))?;
+        }
+        if added.contains(Modifier::BOLD) {
+            queue!(writer, SetAttribute(CAttribute::Bold))?;
+        }
+        if added.contains(Modifier::ITALIC) {
+            queue!(writer, SetAttribute(CAttribute::Italic))?;
+        }
+        if added.contains(Modifier::UNDERLINED) {
+            queue!(writer, SetAttribute(CAttribute::Underlined))?;
+        }
+        if added.contains(Modifier::DIM) {
+            queue!(writer, SetAttribute(CAttribute::Dim))?;
+        }
+        if added.contains(Modifier::CROSSED_OUT) {
+            queue!(writer, SetAttribute(CAttribute::CrossedOut))?;
+        }
+
+        Ok(())
+    }
+}
